sinewave_stress_test.py
Before I start developing GUI for PlainDAQ, I wanted to see how far I can push it. 
In this stress test I tried to use matplotlib's animation functionality, and pushed it to its limit.
The interval of the animation is set 1ms to make it update the display as fast as possible. In each frame a 2000 sample long sinewave with different frequency, amplitude and phase is generated for the animation and there are other stuff that are update for each frame as well.
It can achieve about 60FPS in my computer which more than I need.